Upgrade to agrona 0.5.1

+++ src/main/java/com/lmax/angler/monitoring/network/monitor/socket/udp/UdpSocketMonitor.java
@@ -19,7 +19,7 @@ import java.util.concurrent.atomic.AtomicReference;
  */
 public final class UdpSocketMonitor
 {
-    private static final double AGRONA_DEFAULT_LOAD_FACTOR = 0.67;
+    private static final float AGRONA_DEFAULT_LOAD_FACTOR = 0.67f;
 
     private final Long2ObjectHashMap<UdpBufferStats> monitoredSocketInstances = new Long2ObjectHashMap<>();
     private final AtomicReference<Long2ObjectHashMap<InetSocketAddress>> candidateSockets = new AtomicReference<>(new Long2ObjectHashMap<>());
